+++ packages/hetu_script/lib/interpreter/binding/class_binding.dart
@@ -319,6 +319,42 @@ class HTMathClassBinding extends HTExternalClass {
                 Map<String, dynamic> namedArgs = const {},
                 List<HTType> typeArgs = const []}) =>
             math.Random().nextInt(positionalArgs.first as int);
+      case 'Math.randomBool':
+        return (HTEntity entity,
+                {List<dynamic> positionalArgs = const [],
+                Map<String, dynamic> namedArgs = const {},
+                List<HTType> typeArgs = const []}) =>
+            math.Random().nextBool();
+      case 'Math.randomColorHex':
+        return (HTEntity entity,
+            {List<dynamic> positionalArgs = const [],
+            Map<String, dynamic> namedArgs = const {},
+            List<HTType> typeArgs = const []}) {
+          var prefix = '#';
+          if (namedArgs['hasAlpha']) {
+            prefix += 'ff';
+          }
+          return prefix +
+              (math.Random().nextDouble() * 16777215)
+                  .truncate()
+                  .toRadixString(16)
+                  .padLeft(6, '0');
+        };
+      case 'Math.randomBrightColorHex':
+        return (HTEntity entity,
+            {List<dynamic> positionalArgs = const [],
+            Map<String, dynamic> namedArgs = const {},
+            List<HTType> typeArgs = const []}) {
+          var prefix = '#';
+          if (namedArgs['hasAlpha']) {
+            prefix += 'ff';
+          }
+          return prefix +
+              (math.Random().nextDouble() * 5592405 + 11184810)
+                  .truncate()
+                  .toRadixString(16)
+                  .padLeft(6, '0');
+        };
       case 'Math.sqrt':
         return (HTEntity entity,
                 {List<dynamic> positionalArgs = const [],