+++ packages/hetu_script_dev_tools/test/test.py
@@ -1,4 +1,4 @@
 import os
 
-os.system('dart bin/cli_tool.dart compile example/script/hello.hts')
+os.system('dart bin/cli_tool.dart compile example/script/hello.hts -o example/script/hello.out')
 os.system('dart bin/cli_tool.dart run example/script/hello.out')